@@ -10,29 +10,26 @@
     "googleapis": "^148.0.0",
     "localforage": "^1.10.0",
     "match-sorter": "^6.3.1",
-    "react": "^19.0.0",
-    "react-dom": "^19.0.0",
-    "react-icons": "^5.0.1",
+    "postcss": "^8.5.3",
+    "react": "^19.1.0",
+    "react-dom": "^19.1.0",
+    "react-icons": "^5.5.0",
     "react-pdf": "^9.2.1",
     "react-responsive-carousel": "^3.2.23",
-    "react-router": "^7.2.0",
-    "react-router-dom": "^7.2.0",
-    "react-scripts": "5.0.1",
+    "react-router": "^7.5.1",
+    "react-router-dom": "^7.5.1",
     "sort-by": "^1.2.0",
-    "tailwindcss": "^3.4.1",
-    "web-vitals": "^3.5.1"
+    "tailwindcss": "^4.1.4"
   },
   "scripts": {
     "predeploy": "npm run build",
-    "deploy": "gh-pages -d build",
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject",
+    "deploy": "gh-pages -d dist",
+    "start": "vite",
+    "build": "vite build",
     "storybook": "storybook dev -p 6006",
     "build-storybook": "storybook build",
     "chromatic": "npx chromatic --project-token=chpt_fa61e504e2e2da8",
-    "serve": "serve -s build",
+    "serve": "vite preview",
     "start-server": "node server.js"
   },
   "eslintConfig": {
@@ -58,16 +55,21 @@
     "@storybook/addon-interactions": "^8.0.0",
     "@storybook/addon-links": "^8.0.0",
     "@storybook/blocks": "^8.0.0",
-    "@storybook/preset-create-react-app": "^8.0.0",
+    "@storybook/builder-vite": "^8.2.6",
     "@storybook/react": "^8.0.0",
-    "@storybook/react-webpack5": "^8.0.0",
-    "@storybook/testing-library": "^0.2.2",
+    "@storybook/react-vite": "^8.2.6",
+    "@tailwindcss/postcss": "^4.1.4",
+    "@types/node": "^22.14.1",
+    "@types/react": "^19.1.2",
+    "@types/react-dom": "^19.1.2",
+    "@vitejs/plugin-react": "^4.3.1",
+    "autoprefixer": "^10.4.21",
     "babel-plugin-named-exports-order": "^0.0.2",
     "chromatic": "^10.2.2",
     "gh-pages": "^6.1.1",
     "prop-types": "^15.8.1",
     "storybook": "^8.0.0",
-    "webpack": "^5.89.0"
+    "vite": "^5.4.0"
   },
   "readme": "ERROR: No README data found!",
   "_id": "ahusan.github.io@0.1.0"
