@@ -1,40 +1,8 @@
 /** @type {import('tailwindcss').Config} */
-
-const defaultTheme = require('tailwindcss/defaultTheme');
-
-module.exports = {
-  content: [
-    './src/**/*.{html,js,jsx,ts,tsx}',
-    './pages/**/*.{html,js,jsx,ts,tsx}',
-    './components/**/*.{html,js,jsx,ts,tsx}',
-  ],
+export default {
+  content: ['./src/**/*.{js,jsx,ts,tsx}', './public/index.html'],
   theme: {
-    extend: {
-      fontFamily: {
-        sans: ['Inter var', ...defaultTheme.fontFamily.sans],
-      },
-      animation: {
-        blink: 'blink 1s step-end infinite',
-        float: 'float 3s ease-in-out infinite alternate',
-        'pulse-slow': 'pulse 3s cubic-bezier(0.4, 0, 0.6, 1) infinite',
-      },
-      keyframes: {
-        blink: {
-          '0%, 100%': { opacity: 1 },
-          '50%': { opacity: 0 },
-        },
-        float: {
-          '0%': { transform: 'translateY(0px)' },
-          '100%': { transform: 'translateY(-10px)' },
-        },
-      },
-      backgroundImage: {
-        'gradient-radial': 'radial-gradient(var(--tw-gradient-stops))',
-      },
-      boxShadow: {
-        glow: '0 0 15px 5px rgba(79, 70, 229, 0.3)',
-      },
-    },
+    extend: {},
   },
   plugins: [],
 };
